changed the files and move the mock data to  __mocks__ folder
